feature: refactor (#38)
* dynamic sql mappers

* refactored catalog-count and catalog-count-enum

* modified catalog-count api to be under branch/id/catalog-count

+++ flyway-db-configuration/flyway-6.5.5/sql/V0.0.1__adding_branch.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS `catalog_count`.`branch` (
-  `id`              BIGINT(8) UNSIGNED NOT NULL AUTO_INCREMENT,
+  `id`              INT UNSIGNED NOT NULL AUTO_INCREMENT,
   `address`         VARCHAR(45) NOT NULL,
   `name`            VARCHAR(45) NOT NULL,
   `registration`    DATETIME NOT NULL,
