+++ flyway-db-configuration/flyway-6.5.5/sql/V0.0.2__adding_persona.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS `persona` (
-  `id`              BIGINT(8) UNSIGNED NOT NULL AUTO_INCREMENT,
+  `id`              INT UNSIGNED NOT NULL AUTO_INCREMENT,
   `name`            VARCHAR(64) NOT NULL,
   `lastname`        VARCHAR(64) NOT NULL,
   `registration`    DATETIME NOT NULL,
@@ -8,3 +8,5 @@ CREATE TABLE IF NOT EXISTS `persona` (
 )
   DEFAULT CHARSET =utf8
   COLLATE =utf8_unicode_ci;
+
+  -- insert into persona(name, lastname, registration, status) values('Javier', 'Meza', current_timestamp(), 1)
