+++ flyway-db-configuration/flyway-6.5.5/sql/V0.0.3.0__catalog_count_enum.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS `catalog_count_enum` (
-  `id`                  BIGINT(8) UNSIGNED NOT NULL AUTO_INCREMENT,
+  `id`                  INT UNSIGNED NOT NULL AUTO_INCREMENT,
   `identifier`          VARCHAR(64)        NOT NULL,
   `family`              VARCHAR(64)        NULL,
   `name`                VARCHAR(64)        NOT NULL,
